Major refactoring of rendering/scene, work in progress

+++ trunk/JonsEngine/include/Core/Types.h
@@ -18,8 +18,9 @@ namespace JonsEngine
         TEXTURE_TYPE_NORMAL
     };
     
-    typedef uint32_t MeshID;
-    typedef uint32_t TextureID;
+    // TODO: move mesh/texture id definitions
+    typedef uint64_t MeshID;
+    typedef uint64_t TextureID;
 
     const MeshID INVALID_MESH_ID = 0;
     const TextureID INVALID_TEXTURE_ID = 0;
